remove the ipsec interception code.
i wrote it cos bridge(4) allowed ipsec to apply policy to ip packets
as they traversed the bridge, but the locking and packet gymnastics
meant it was never enabled. i'm no longer interested in making veb
feature compatible with bridge(4), so this code can go. it's in the
attic if it's needed in the future though.

no functional change.