Add a makefile to link rpki-client against openssl
If openssl and libretls are installed, this provides an easy way of
building an rpki-client binary using OpenSSL. On the one hand this
will ensure that we can build against -current OpenSSL and provides
an easy way to creating such a binary for testing purposes without
needing to deal with autoconf. I also plan to revamp our tests to
use file mode instead of the nasty reacharounds, so that we can use
the same set of tests against both libressl and openssl.

discussed with bluhm and job who both encouraged me to commit this.

Now I need a shower.