First draft of a PRUNING subsection in the manual page which people will read when sysupgrade prevents a dangerously-failing upgrade because /usr is too full.
This explains some of the history and reasons why /usr is too full,
and also explains that sysupgrade / installer cannot delete any
detritus because it cannot identify it.  More imporatntly, users also
will have a hard time identifying the detritus -- most of it is older
lib*.so.* libraries but there's no reasonable mechanism to traverse all
possible filesystems and ensure no binaries remain which use those
libraries.

rather than telling people hacky choices that might get them through this
sysupgrade, and fail next time, we choose to provide no hacky advice.

A bigger /usr is their best choice, which implies that it is time for a reinstall.

ok florian, sthen