vmboot: A tiny kernel for booting SEV VMs
When using SEV-enabled VMs, we let vmd(8) launch the kernel directly.
With this, we do not have to trust a firmware like OVMF.  However,
with this approach, sysupgrade(8) can not be used as the kernel
resides outside the VM in the host file system.  Thus maintainig
SEV-enabled VMs is hard.

With vmboot, we use a small ramdisk kernel that runs the actual
kernel using the kexec facility.  vmboot can be launched directly
by vmd(8).  The actual kernel is loaded from the disk image of the
SEV-enabled VM.  With this, sysupgrade(8) can be used, and there
is no need for additional firmware.

This approach is similar to vmboot for powerpc64 and octeon.

Discussed with bluhm@ for a while.

ok mlarkin@