Some AMD microcode now has a minimum version for updating with an MSR. AMD-SB-7033 claims a protection fault will be generated if not at that level.
The microcode containers can now also have multiple microcode patches
for a given family-model-stepping combination.  One with a minimum version,
and one without.

Handle multiple patches and check minimum version using part of the
container header.  Limit this check to family 19h and above.  The header
is not publically documented and this part seems to have another use on
earlier CPU families.