move aggr/trunk input processing into ether_input.
previously it replaced the ifp->if_input function pointer on ethernet
interfaces, which is always ether_input, to intercept packets. this
makes it work the same as bridges (bridge/veb/tpmr) for intercepting
and filtering packets now.

this makes it (more) possible to call ether_input without netlock
(which serialised the modification/use of if_input), and makes it
possible for ethernet interfaces with custom if_input functions
that eventually call ether_input to work as trunk/aggr ports.