rpki-client: convert mft.c to opaque ASN1_STRING
Mostly more of the same. In addition the comparison functions can use
ASN1_STRING_cmp() instead of doing most of it by hand. The additional
type check this introduces is fine since the ASN.1 template ensures
that. Annotate a spot where we fail to ensure that the unused bits
octet is zero, thus libcrypto's misdesign introduces a few bits of
malleability for every hash that ends in zero bits. This will be easier
to fix later.

ok claudio