move the tcp code in the pf_test pd.virtual_proto switch statement up
there's a switch statement in pf_test() that handles pd.virtual_proto.
however, the very first line of the code in the default case is
"if (pd.virtual_proto == IPPROTO_TCP)", so we can hoist that up
above the default case and then have it fall through.

no functional change.

ok sashan@