don't mess around with m->m_pkthdr.pf.inp and inp links in pf_test_state.
pf_test_state copies the inp pointer from the mbuf, and then calls
pf_mbuf_link_inpcb to set it again. nothing between these statements
clears the mbuf inp pointer. this only works if the mbuf inp pointer
is always NULL because pf_mbuf_link_inpcb asserts that the mbuf inp
pointer is NULL before setting it.

this relies on the inp pointer not being set for incoming packets,
and the inp pointer getting cleared by pf_find_state for outgoing
packets.

at worst it's a hazard for the future, at best it's a waste of time.
let's just get rid of it.

ok sashan@